Allow using an offset for the issue numbers
This is helpful when importing issues into an existing repository with
some existing PRs, as they take consume numbers in the same namespace as
the issues, making it impossible to preserve (all) the existing issue
numbers. Using this option it's at least possible to start importing
from some round number, e.g. 100, or 1000, so that the old ticket N is
now mapped to 100+N or 1000+N.